Add clock stretching support to I2C master
Add clock stretching support to the distributed I2C master
implementation. This commit also sorts out the timing so that if the
clock is not being stretched then the data will be transferred at the
specified rate - previously the actual rate was well under the specified
rate (for example if you asked for 400Kpbs, you got ~320Kbps).