Improve rst substitution to superscript I2C
The new substitution renders correctly when viewing the README rst on
github.com. It is now used in all titles, figure captions etc.

Pattern has been changed to uppercase so that it is output as "I2C" if
used in places where xdoc is unable to render superscript, as well as
reading better when viewing the 'raw' rst.