Fix i2c_master_reg tests
reg_ops_nack.expect was incorrect.  Based on git history, the expected
value from a NACK occurring before data when read_reg16_addr8 was called
was resulting in the test failing.  The old behavior was undefined.  The
new behavior for all register read variations is to return 0 on this
failure.